+++ spec/cov_loupe/cli/cli_enumerated_options_spec.rb
@@ -30,6 +30,10 @@ RSpec.describe 'CLI enumerated option parsing' do
       { argv: %w[--source full summary lib/foo.rb], accessor: :source_mode, expected: :full },
       { argv: %w[--source uncovered summary lib/foo.rb], accessor: :source_mode,
         expected: :uncovered },
+      { argv: %w[--source n summary lib/foo.rb], accessor: :source_mode, expected: nil },
+      { argv: %w[--source none summary lib/foo.rb], accessor: :source_mode, expected: nil },
+      { argv: %w[-s n summary lib/foo.rb], accessor: :source_mode, expected: nil },
+      { argv: %w[-s none summary lib/foo.rb], accessor: :source_mode, expected: nil },
 
       { argv: %w[-S yes list], accessor: :raise_on_stale, expected: true },
       { argv: %w[--raise-on-stale yes list], accessor: :raise_on_stale, expected: true },